Separate limits for individual use, alter recommended to be the most common use
