+++ eslint.config.mjs
@@ -2,14 +2,14 @@ import tseslint from 'typescript-eslint'
 
 import af from './src/index.js'
 
-// npx @eslint/config-inspector
+/**
+ * Project eslint configuration.
+ * 
+ * View config with `npx @eslint/config-inspector`
+ */
 export default tseslint.config(
   {
     name: 'project',
-    extends: [...af.configs.recommended, ...af.rulesets.jsOnly],
-  },
-  {
-    files: ['./scripts/**/*.js'],
-    extends: [...af.rulesets.jsOnlyScripts],
+    extends: [...af.configs.recommended, ...af.limits.jsOnly],
   },
 )
