+++ scripts/clean-yarn.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import { $ } from 'execa'
 import { rimraf as r } from 'rimraf'
