+++ scripts/reset.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import { dirname } from 'node:path'
 import { fileURLToPath } from 'node:url'
