+++ src/configs/index.js
@@ -1,3 +1,4 @@
+import jsOnlyScripts from '../limits/jsOnlyScripts.js'
 import js from './js.js'
 import json from './json.js'
 import markdown from './markdown.js'
@@ -6,6 +7,7 @@ const configs = {
   js,
   json,
   markdown,
-  recommended: [...js, ...json, ...markdown],
+  recommended: [...js, ...jsOnlyScripts, ...json, ...markdown],
+  other: [...markdown, ...json],
 }
 export default configs
