+++ src/configs/json.js
@@ -3,7 +3,11 @@ import tseslint from 'typescript-eslint'
 
 import { BUILD_IGNORES, JSON_FILES } from '../constants.js'
 
-// npx @eslint/config-inspector
+/**
+ * Configuration preset for json files
+ *
+ * View configuration with `npx @eslint/config-inspector`
+ */
 const configs = tseslint.config(
   {
     name: 'alienfast-json-ignores',
