+++ src/limits/jsOnly.js
@@ -3,10 +3,12 @@ import tseslint from 'typescript-eslint'
 /**
  * Turn off rules not necessary for js only files.
  *
+ * Do not extend configs, it will alter files/ignores behavior.
+ *
  * View config with `npx @eslint/config-inspector`
  */
 const configs = tseslint.config({
-  name: 'alienfast-ruleset-js-only',
+  name: 'alienfast-js-only',
   rules: {
     '@typescript-eslint/no-unsafe-argument': 'off',
     '@typescript-eslint/no-unsafe-member-access': 'off',
