+++ src/limits/jsOnlyScripts.js
@@ -1,15 +1,19 @@
 import tseslint from 'typescript-eslint'
 
+import { SCRIPTS } from '../constants.js'
 import jsOnly from './jsOnly.js'
 
 /**
  * Turn off rules not necessary for scripts
  *
+ * Do not extend configs, it will alter files/ignores behavior.
+ *
  * View config with `npx @eslint/config-inspector`
  */
 const configs = tseslint.config({
-  name: 'alienfast-ruleset-js-only-scripts',
+  name: 'alienfast-js-only-scripts',
   extends: [...jsOnly],
+  files: SCRIPTS,
   rules: {
     'no-console': 'off',
   },
